Revert "Apply optimistic caching to executeField, too."
Since executeField is typically a cheap operation, the benefits of caching
have to outweigh the overhead of caching. Empirically, that balance does
not favor caching the executeField function.

If a field has a selection set associated with it, executeSelectionSet
will be called, and the result will be cached. In other words, removing
caching for executeField primarily eliminates the caching overhead for
fields that just have scalar values.